WIP: Cleaned-up interfaces round 1

+++ org.knime.storage/src/main/java/org/knime/core/data/table/column/NativeType.java
@@ -1,5 +1,5 @@
 package org.knime.core.data.table.column;
 
 public enum NativeType {
-	STRING, BOOLEAN, LONG;
+	STRING, BOOLEAN, DOUBLE;
 }
